+++ src/components/CardHome/styled.js
@@ -1,9 +1,6 @@
-import React from 'react'
-import styled from 'styled-components'
-import { useHistory } from 'react-router-dom'
-import { goToRestaurant } from '../routes/coordinator'
+import styled from "styled-components"
 
-const MainCard = styled.div`
+export const MainCard = styled.div`
     width: 20.5rem;
     height: 11.75rem;
     padding: 0 0 1rem;
@@ -11,14 +8,14 @@ const MainCard = styled.div`
     border: solid 1px #b8b8b8;
 `
 
-const ImgCard = styled.img`
+export const ImgCard = styled.img`
     width: 20.5rem;
     height: 7.5rem;
     margin: 0 0 0.75rem;
     object-fit: contain;
 `
 
-const RestauranteNome = styled.span`
+export const RestauranteNome = styled.span`
     width: 18.5rem;
     height: 1.125rem;
     margin: 0.75rem 1rem 0.25rem;
@@ -32,7 +29,7 @@ const RestauranteNome = styled.span`
     color: #5cb646;
 `
 
-const RestauranteTempo = styled.span`
+export const RestauranteTempo = styled.span`
     width: 9.25rem;
     height: 1.125rem;
     margin: 0.25rem 0.5rem 0 1rem;
@@ -46,7 +43,7 @@ const RestauranteTempo = styled.span`
     color: #b8b8b8;
 `
 
-const RestauranteFrete = styled.span`
+export const RestauranteFrete = styled.span`
     width: 8.75rem;
     height: 1.125rem;
     margin: 0.25rem 1rem 0 0.5rem;
@@ -61,12 +58,12 @@ const RestauranteFrete = styled.span`
     color: #b8b8b8;
 `
 
-const ContainerFrete = styled.div`
+export const ContainerFrete = styled.div`
     display: flex;
     justify-content: space-between;
 `
 
-const RestauranteCard = styled.div`
+export const RestauranteCard = styled.div`
     width: 22.5rem;
     height: 12.25rem;
     margin: 3.125rem 0 0;
@@ -77,27 +74,8 @@ const RestauranteCard = styled.div`
     justify-content: center;
 `
 
-const ContainerNome = styled.div`
+export const ContainerNome = styled.div`
     display: flex;
     flex-direction: column;
     justify-content: space-around;
 `
-function CardHome (props) {
-    const history = useHistory()
-    return(
-        <RestauranteCard>
-            <MainCard onClick={() => goToRestaurant(history, props.restaurante.id)}>
-                <ImgCard src={props.restaurante.logoUrl} />
-                <ContainerNome>
-                    <RestauranteNome>{props.restaurante.name}</RestauranteNome>
-                    <ContainerFrete>
-                        <RestauranteTempo>{props.restaurante.deliveryTime} min</RestauranteTempo>
-                        <RestauranteFrete>Frete R${props.restaurante.shipping},00</RestauranteFrete>
-                    </ContainerFrete>
-                </ContainerNome>
-            </MainCard>
-        </RestauranteCard>
-    )
-}
-
-export default CardHome
