+++ src/constants/urls.js
@@ -1,2 +1,1 @@
-
 export const BASE_URL = "https://us-central1-missao-newton.cloudfunctions.net/futureEatsA"
